+++ src/main/java/racingcar/Cars.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import java.util.ArrayList;
 import java.util.List;
