+++ src/main/java/racingcar/README.md
@@ -1,4 +1,15 @@
+# Step 3
+
 - [x] 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
 - [x] 사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
 - [x] 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
 - [x] 자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.
+
+# Step 4
+
+- [ ] 자동차 이름 프로퍼티
+- [ ] 자동차 이름 validation
+- [ ] 출력 시 자동차 이름 출력
+- [ ] 자동차 이름 input 쉼표 구분
+- [ ] 자동차 이름 input validation
+- [ ] 우승자 noti 추가 (복수 우승 가능)
