+++ src/main/java/racingcar/Racer.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 public class Racer {
