+++ src/test/java/racingcar/CarTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
