+++ src/test/java/racingcar/RaceEvaluatorTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
